Moved WAVEDescriptor into ama.py
